requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect==1.2.0
distro==1.9.0
dj-database-url==2.2.0
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jieba3k==0.35.1
jiter==0.5.0
joblib==1.4.2
kombu==5.4.0
lxml==5.3.0
lxml_html_clean==0.2.0
newspaper3k==0.2.8
nltk==3.9.1
openai==1.42.0
pillow==10.4.0
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
gTTS==2.5.3
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
pydub==0.25.1
requests==2.32.3
tqdm==4.66.5
